requirements.txt
# ── Core AI & data ────────────────────────────────────────────────────────────
anthropic
requests
python-dotenv
yfinance
ta
pyyaml
numpy
pandas

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit
streamlit-autorefresh
plotly

# ── Scheduler ─────────────────────────────────────────────────────────────────
schedule

# ── Database ──────────────────────────────────────────────────────────────────
# psycopg2-binary is only needed in production (PostgreSQL).
# For local SQLite development it is optional.
psycopg2-binary==2.9.9

# ── Broker / execution ──────────────────────────────────────────────────────
alpaca-trade-api
alpha_vantage==2.3.1

# ── Macro data ───────────────────────────────────────────────────────────────
fredapi

# ── Health check server ───────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Testing ───────────────────────────────────────────────────────────────────
pytest
pytest-cov
pytest-mock
responses
